requirements.txt
blinker==1.7.0
cachetools==5.3.2
click==8.1.7
Flask==3.0.1
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gspread==5.12.4
itsdangerous==2.1.2
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests-oauthlib==1.3.1
rsa==4.9
tabulate==0.9.0
termcolor==2.4.0
Werkzeug==3.0.1